merge azure-powershell/master

+++ src/Compute/Compute/Extension/CustomScript/RemoveAzureVMCustomScriptExtensionCommand.cs
@@ -57,23 +57,38 @@ namespace Microsoft.Azure.Commands.Compute
         [ValidateNotNullOrEmpty]
         public SwitchParameter Force { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Starts the operation and returns immediately, before the operation is completed. In order to determine if the operation has sucessufuly been completed, use some other mechanism.")]
+        public SwitchParameter NoWait { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
 
             ExecuteClientAction(() =>
             {
-                if (this.ShouldProcess(VMName, Properties.Resources.RemoveScriptExtensionAction) 
-                && (this.Force.IsPresent 
-                || this.ShouldContinue(Properties.Resources.VirtualMachineExtensionRemovalConfirmation, 
+                if (this.ShouldProcess(VMName, Properties.Resources.RemoveScriptExtensionAction)
+                && (this.Force.IsPresent
+                || this.ShouldContinue(Properties.Resources.VirtualMachineExtensionRemovalConfirmation,
                             Properties.Resources.VirtualMachineExtensionRemovalCaption)))
                 {
-                    var op = this.VirtualMachineExtensionClient.DeleteWithHttpMessagesAsync(
-                        this.ResourceGroupName,
-                        this.VMName,
-                        this.Name).GetAwaiter().GetResult();
-                    var result = ComputeAutoMapperProfile.Mapper.Map<PSAzureOperationResponse>(op);
-                    WriteObject(result);
+                    if (NoWait.IsPresent)
+                    {
+                        var op = this.VirtualMachineExtensionClient.BeginDeleteWithHttpMessagesAsync(
+                            this.ResourceGroupName,
+                            this.VMName,
+                            this.Name).GetAwaiter().GetResult();
+                        var result = ComputeAutoMapperProfile.Mapper.Map<PSAzureOperationResponse>(op);
+                        WriteObject(result);
+                    }
+                    else
+                    {
+                        var op = this.VirtualMachineExtensionClient.DeleteWithHttpMessagesAsync(
+                            this.ResourceGroupName,
+                            this.VMName,
+                            this.Name).GetAwaiter().GetResult();
+                        var result = ComputeAutoMapperProfile.Mapper.Map<PSAzureOperationResponse>(op);
+                        WriteObject(result);
+                    }
                 }
             });
         }
